fix texture y axis

+++ src/cube3d.h
@@ -6,7 +6,7 @@
 /*   By: capapes <capapes@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/31 17:21:13 by capapes           #+#    #+#             */
-/*   Updated: 2025/03/05 19:00:09 by capapes          ###   ########.fr       */
+/*   Updated: 2025/03/07 13:10:45 by capapes          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -283,7 +283,7 @@ double					get_side_len(t_vector a1, t_vector a2, t_vector tan,
 double					radian_overflow(double angle);
 t_texture				get_texture(int set, double origin_x_percent);
 uint32_t				get_pixel_info(mlx_image_t *texture,
-							uint16_t pixel_index);
+							uint32_t pixel_index);
 uint32_t				get_texture_color(t_texture texture);
 
 // ----------------------------[DELETE BEFORE SUBMIT]---------------
